Make twig extensions lazy
