+++ src/Twig/RolesMatrixExtension.php
@@ -13,74 +13,16 @@ declare(strict_types=1);
 
 namespace Nucleos\UserAdminBundle\Twig;
 
-use Nucleos\UserAdminBundle\Security\RolesBuilder\MatrixRolesBuilderInterface;
-use Symfony\Component\Form\FormView;
-use Twig\Environment;
 use Twig\Extension\AbstractExtension;
 use Twig\TwigFunction;
 
 final class RolesMatrixExtension extends AbstractExtension
 {
-    private MatrixRolesBuilderInterface $rolesBuilder;
-
-    public function __construct(MatrixRolesBuilderInterface $rolesBuilder)
-    {
-        $this->rolesBuilder = $rolesBuilder;
-    }
-
-    /**
-     * @return TwigFunction[]
-     */
     public function getFunctions(): array
     {
         return [
-            new TwigFunction('renderMatrix', [$this, 'renderMatrix'], ['needs_environment' => true]),
-            new TwigFunction('renderRolesList', [$this, 'renderRolesList'], ['needs_environment' => true]),
+            new TwigFunction('renderMatrix', [RolesMatrixRuntime::class, 'renderMatrix'], ['needs_environment' => true]),
+            new TwigFunction('renderRolesList', [RolesMatrixRuntime::class, 'renderRolesList'], ['needs_environment' => true]),
         ];
-    }
-
-    public function renderRolesList(Environment $environment, FormView $form): string
-    {
-        $roles = $this->rolesBuilder->getRoles();
-        foreach ($roles as $role => $attributes) {
-            if (isset($attributes['admin_label'])) {
-                unset($roles[$role]);
-
-                continue;
-            }
-
-            $roles[$role] = $attributes;
-            foreach ($form->getIterator() as $child) {
-                if ($child->vars['value'] === $role) {
-                    $roles[$role]['form'] = $child;
-                }
-            }
-        }
-
-        return $environment->render('@NucleosUserAdmin/Form/roles_matrix_list.html.twig', [
-            'roles' => $roles,
-        ]);
-    }
-
-    public function renderMatrix(Environment $environment, FormView $form): string
-    {
-        $groupedRoles = [];
-        foreach ($this->rolesBuilder->getRoles() as $role => $attributes) {
-            if (!isset($attributes['admin_label'])) {
-                continue;
-            }
-
-            $groupedRoles[$attributes['admin_label']][$role] = $attributes;
-            foreach ($form->getIterator() as $child) {
-                if ($child->vars['value'] === $role) {
-                    $groupedRoles[$attributes['admin_label']][$role]['form'] = $child;
-                }
-            }
-        }
-
-        return $environment->render('@NucleosUserAdmin/Form/roles_matrix.html.twig', [
-            'grouped_roles'     => $groupedRoles,
-            'permission_labels' => $this->rolesBuilder->getPermissionLabels(),
-        ]);
     }
 }
